changed type from int to float

+++ 11JulyTask1.py
@@ -1,3 +1,3 @@
-start = int(input("Number of hours: "))
+start = float(input("Number of hours: "))
 answer = start*60
 print(str(start) + " hours is equal to " + str(answer) + " minutes ")
